Source ~/.profile instead of ~/.bash_aliases
The rationale for sourcing ~/.profile instead of ~/.bash_aliases is to
export the $XDG_* variables, in particular $XDG_CONFIG_HOME, which is
needed by the `dotgit` alias. This also has the side effect of sourcing
~/.bash_aliases, which defines said alias.

The $HOME/.config and $HOME/.cache directories are also checked to make
sure they exist before the dotfiles repo is cloned, otherwise the script
will break.