Require script to be sourced from current shell
Previously, dotfiles-install.sh was executed from the user's current
shell. As a consequence, the command '. ~/.profile' did not modify the
parent shell's environment. Thus, the dotfiles-install.sh script is now
intended to be sourced.

The '.sh' prefix has been removed, and the script no longer deletes
itself on (successful) completion.