Stop cp from prompting overwrites, remove traps
It seems that on many systems, cp is aliased to 'cp -i', which prompts
before overwrites. A leading '\' on the command will prevent the alias
from being called, stopping this (dotfiles-install is meant to not
require any interaction).

The traps have also been taken out, as messing with traps in the user's
working shell can cause some confusion if something goes wrong with the
script's execution.