Add check to make sure script is sourced
This script is only meant to be sourced; the messy function
'_check_if_sourced' has been added to provide this check, in a
POSIX-compliant (but not fool-proof) way.

The script now returns on error in the top-level scope, as returning
from a function ('_error', in this case) does not stop the calling shell
from continuing to execute lines of code.